horror_bot/requirements.txt
discord.py
python-dotenv
aiosqlite
llama-cpp-python
huggingface-hub